Split out the inspector
