+++ lib/ensureAfterHookIsRegistered.js
@@ -1,14 +1,16 @@
 /*global after*/
+const ruleFixer = require('eslint/lib/util/rule-fixer');
+const Traverser = require('eslint/lib/util/traverser');
+const isSimpleObjectTree = require('./isSimpleObjectTree');
 const memoizeSync = require('memoizesync');
 const detectIndent = require('detect-indent');
 
 const indentString = require('./indentString');
-function stringify(obj, indentationWidth, expectForRendering) {
+function stringify(obj, indentationWidth, inspect) {
   if (obj.includes('\n')) {
     return `expect.unindent\`${indentString(obj, indentationWidth)}\``;
   } else {
-    expectForRendering.output.indentationWidth = indentationWidth;
-    return expectForRendering.inspect(obj).toString('text');
+    return inspect(obj, indentationWidth);
   }
 }
 
@@ -140,7 +142,7 @@ function ensureAfterBlockIsRegistered(topLevelFixes) {
       columnNumber,
       status,
       subject,
-      expectForRendering,
+      inspect,
       assertionName
     } of topLevelFixes) {
       const sourceCode = getSourceCode(fileName);
@@ -177,9 +179,6 @@ function ensureAfterBlockIsRegistered(topLevelFixes) {
           return '';
         }
       }
-      const ruleFixer = require('eslint/lib/util/rule-fixer');
-      const Traverser = require('eslint/lib/util/traverser');
-      const isSimpleObjectTree = require('./isSimpleObjectTree');
 
       new Traverser().traverse(sourceCode.ast, {
         enter(node, parent) {
@@ -211,22 +210,19 @@ function ensureAfterBlockIsRegistered(topLevelFixes) {
               stringifiedSubject = stringify(
                 subject,
                 indentationWidth,
-                expectForRendering
+                inspect
               ).replace(/\n^(?=[^\n])/gm, `\n${indent}`);
             } else if (
-              isSimpleObjectTree(subject, expectForRendering) &&
+              isSimpleObjectTree(subject) &&
               assertionName === 'to equal snapshot'
             ) {
-              expectForRendering.output.indentationWidth = indentationWidth;
-              stringifiedSubject = expectForRendering
-                .inspect(subject)
-                .toString('text');
+              stringifiedSubject = inspect(subject, indentationWidth);
             } else {
               newAssertionName = 'to inspect as snapshot';
               stringifiedSubject = stringify(
-                expectForRendering.inspect(subject).toString('text'),
+                inspect(subject, indentationWidth),
                 indentationWidth,
-                expectForRendering
+                inspect
               ).replace(/\n^(?=[^\n])/gm, `\n${indent}`);
             }
             if (status === 'missing') {
@@ -348,7 +344,7 @@ function ensureAfterBlockIsRegistered(topLevelFixes) {
 
         const Promise = require('bluebird');
         const fs = Promise.promisifyAll(require('fs'));
-        console.log(fileNames.length);
+
         if (fileNames.length > 0) {
           const results = await Promise.map(
             fileNames,
