+++ lib/isSimpleObjectTree.js
@@ -1,4 +1,6 @@
-function isSimpleObjectTree(obj, expect) {
+const expect = require('unexpected');
+
+function isSimpleObjectTree(obj) {
   const seen = new Set();
   return (function isSimple(obj) {
     if (seen.has(obj)) {
@@ -29,7 +31,6 @@ function isSimpleObjectTree(obj, expect) {
           'undefined',
           'string',
           'Buffer',
-          'infiniteBuffer', // expectForRendering's hack
           'date'
         ].includes(type.name);
       }
