+++ lib/unexpected-snapshot-check.js
@@ -1,7 +1,11 @@
+const createInspector = require('./createInspector');
+
 module.exports = {
   name: 'unexpected-snapshot',
   version: require('../package.json').version,
   installInto(expect) {
+    const inspect = createInspector(expect);
+
     expect.unindent = require('@gustavnikolaj/string-utils').deindent;
 
     const snapshotMessage =
@@ -46,7 +50,7 @@ module.exports = {
       '<any> to inspect as snapshot <string>',
       (expect, subject, value) => {
         withSnapshotMessageAppended(expect, () => {
-          expect(expect.inspect(subject).toString('text'), 'to equal', value);
+          expect(inspect(subject), 'to equal', value);
         });
       }
     );
