+++ lib/unexpected-snapshot-update.js
@@ -1,4 +1,5 @@
-const _ = require('lodash');
+const createInspector = require('./createInspector');
+const ensureAfterHookIsRegistered = require('./ensureAfterHookIsRegistered');
 
 function extractCallerLocationFromStack(stack) {
   let topFrame = stack
@@ -26,81 +27,7 @@ module.exports = {
   name: 'unexpected-snapshot',
   version: require('../package.json').version,
   installInto(expect) {
-    const expectForRendering = expect.child();
-
-    if (typeof Buffer === 'function') {
-      expectForRendering.addType({
-        name: 'infiniteBuffer',
-        base: 'Buffer',
-        identify(obj) {
-          return this.baseType.identify(obj);
-        },
-        inspect(value, depth, output, inspect) {
-          if (value.length > 32) {
-            return output.code(
-              `Buffer.from('${value.toString('base64')}', 'base64')`,
-              'javascript'
-            );
-          } else {
-            return this.baseType.inspect.call(
-              this,
-              value,
-              depth,
-              output,
-              inspect
-            );
-          }
-        },
-        prefix(output) {
-          return output.code('Buffer.from([', 'javascript');
-        },
-        suffix(output) {
-          return output.code('])', 'javascript');
-        },
-        hexDumpWidth: Infinity // Prevents Buffer instances > 16 bytes from being truncated
-      });
-    }
-
-    expectForRendering.addType({
-      base: 'Error',
-      name: 'overriddenError',
-      identify(obj) {
-        return this.baseType.identify(obj);
-      },
-      inspect(value, depth, output, inspect) {
-        var obj = _.extend({}, value);
-
-        var keys = Object.keys(obj);
-        if (keys.length === 0) {
-          output
-            .text('new Error(')
-            .append(inspect(value.message || ''))
-            .text(')');
-        } else {
-          output
-            .text('(function () {')
-            .text(`var err = new ${value.constructor.name || 'Error'}(`)
-            .append(inspect(value.message || ''))
-            .text(');');
-          keys.forEach(function(key, i) {
-            output.sp();
-            if (/^[a-z$_][a-z0-9$_]*$/i.test(key)) {
-              output.text(`err.${key}`);
-            } else {
-              output
-                .text('err[')
-                .append(inspect(key))
-                .text(']');
-            }
-            output
-              .text(' = ')
-              .append(inspect(obj[key]))
-              .text(';');
-          });
-          output.sp().text('return err;}())');
-        }
-      }
-    });
+    const inspect = createInspector(expect);
 
     const symbol = Symbol('unexpectedSnapshot');
 
@@ -116,7 +43,7 @@ module.exports = {
             assertionName: expect.testDescription,
             status: 'missing'
           });
-          require('./ensureAfterHookIsRegistered')(topLevelFixes);
+          ensureAfterHookIsRegistered(topLevelFixes);
           expect.fail();
         } else {
           throw new Error(
@@ -134,7 +61,7 @@ module.exports = {
       (expect, subject, ...args) => {
         const subjectForAssertion =
           expect.testDescription === 'to inspect as snapshot'
-            ? expectForRendering.inspect(subject).toString('text')
+            ? inspect(subject)
             : subject;
 
         return expect.withError(
@@ -148,7 +75,7 @@ module.exports = {
               assertionName: expect.testDescription,
               status: 'mismatch'
             });
-            require('./ensureAfterHookIsRegistered')(topLevelFixes);
+            ensureAfterHookIsRegistered(topLevelFixes);
             throw err;
           }
         );
@@ -165,7 +92,8 @@ module.exports = {
           } catch (e) {
             context[symbol] = {
               ...extractCallerLocationFromStack(e.stack),
-              expectForRendering
+              inspect,
+              expect
             };
           }
         }
