+++ package.json
@@ -28,7 +28,6 @@
     "eslint": "^5.10.0",
     "espree": "^5.0.0",
     "is-node": "^1.0.2",
-    "lodash": "^4.17.11",
     "memoizesync": "^1.1.1"
   },
   "devDependencies": {
